PROBLEM: -1 imlicitly converted to UINT64_MAX is used for timestamp and it's expected that receiver translates this to current time

+++ src/metriclist.h
@@ -115,7 +115,7 @@ public:
 private:
 
     // Maximum livetime of metric
-    int _maxLiveTime = 5 * 60;
+    uint64_t _maxLiveTime = 5 * 60;
 
     // Metric list <topic, Metric>
     std::map <std::string, MetricInfo> _knownMetrics;
